RatioService

+++ src/main/java/com/noor/wrapper/MedicalPerMonthDTO.java
@@ -5,6 +5,6 @@ import java.io.Serializable;
 public record MedicalPerMonthDTO  (
     Long serviceID,
     String serviceName,
-    int medicalCount) implements Serializable {
+    Long medicalCount) implements Serializable {
 
 }
